remove local file writer from file_writers.py
